+++ intelmq/bin/intelmqctl.py
@@ -1836,7 +1836,7 @@ Get some debugging output on the settings and the enviroment (to be extended):
 
         output = {}
         if sections is None or 'paths' in sections:
-            output['paths'] = []
+            output['paths'] = {}
             variables = globals()
             if RETURN_TYPE == 'text':
                 print('Paths:')
@@ -1845,18 +1845,18 @@ Get some debugging output on the settings and the enviroment (to be extended):
                          'RUNTIME_CONF_FILE', 'VAR_RUN_PATH', 'STATE_FILE_PATH',
                          'DEFAULT_LOGGING_PATH', '__file__',
                          'CONFIG_DIR', 'ROOT_DIR'):
-                output['paths'].append((path, variables[path]))
+                output['paths'][path] = variables[path]
                 if RETURN_TYPE == 'text':
-                    print('%s: %r' % output['paths'][-1])
+                    print('%s: %r' % (path, variables[path]))
         if sections is None or 'environment_variables' in sections:
-            output['environment_variables'] = []
+            output['environment_variables'] = {}
             if RETURN_TYPE == 'text':
                 print('Environment variables:')
             for variable in ('INTELMQ_ROOT_DIR', 'INTELMQ_PATHS_NO_OPT',
                              'INTELMQ_PATHS_OPT', 'INTELMQ_MANAGER_CONTROLLER_CMD'):
-                output['environment_variables'].append((variable, os.getenv(variable)))
+                output['environment_variables'][variable] = os.getenv(variable)
                 if RETURN_TYPE == 'text':
-                    print('%s: %r' % output['environment_variables'][-1])
+                    print('%s: %r' % (variable, os.getenv(variable)))
         return 0, output